requirements.txt
requests
pydantic
pydub
speechrecognition
loguru
pydantic[dotenv]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib